natlab: Improve imports structure to avoid cycling
