+++ nat-lab/tests/test_client_basic_stun.py
@@ -1,7 +1,8 @@
 import config
 import pytest
 from utils import stun
-from utils.connection_util import ConnectionTag, new_connection_by_tag
+from utils.connection import ConnectionTag
+from utils.connection_util import new_connection_by_tag
 
 
 @pytest.mark.asyncio
