+++ nat-lab/tests/test_dns_through_exit.py
@@ -5,7 +5,8 @@ from contextlib import AsyncExitStack
 from helpers import setup_api, setup_mesh_nodes, SetupParameters
 from typing import List, Tuple
 from utils.bindings import default_features, TelioAdapterType
-from utils.connection_util import generate_connection_tracker_config, ConnectionTag
+from utils.connection import ConnectionTag
+from utils.connection_util import generate_connection_tracker_config
 from utils.dns import query_dns
 from utils.router import IPStack
