+++ nat-lab/tests/test_mesh_exit_through_peer.py
@@ -13,9 +13,9 @@ from utils.bindings import (
     RelayState,
     TelioAdapterType,
 )
+from utils.connection import ConnectionTag
 from utils.connection_util import (
     generate_connection_tracker_config,
-    ConnectionTag,
     new_connection_with_conn_tracker,
 )
 from utils.ping import ping
