+++ nat-lab/tests/test_mesh_plus_vpn.py
@@ -15,9 +15,9 @@ from utils.bindings import (
     RelayState,
     NodeState,
 )
+from utils.connection import ConnectionTag
 from utils.connection_util import (
     generate_connection_tracker_config,
-    ConnectionTag,
     new_connection_with_conn_tracker,
 )
 from utils.ping import ping
