+++ nat-lab/tests/test_netcat.py
@@ -2,7 +2,8 @@ import asyncio
 import pytest
 from contextlib import AsyncExitStack
 from helpers import SetupParameters, setup_environment
-from utils.connection_util import ConnectionTag, new_connection_by_tag
+from utils.connection import ConnectionTag
+from utils.connection_util import new_connection_by_tag
 from utils.netcat import NetCatServer, NetCatClient
 
 TEST_STRING = "test_data"
