+++ nat-lab/tests/test_node_state_flickering.py
@@ -12,7 +12,8 @@ from utils.bindings import (
     NodeState,
     RelayState,
 )
-from utils.connection_util import generate_connection_tracker_config, ConnectionTag
+from utils.connection import ConnectionTag
+from utils.connection_util import generate_connection_tracker_config
 
 
 @pytest.mark.asyncio
