+++ nat-lab/tests/test_process_read_loop.py
@@ -1,6 +1,7 @@
 import pytest
 from contextlib import AsyncExitStack
-from utils.connection_util import ConnectionTag, new_connection_by_tag
+from utils.connection import ConnectionTag
+from utils.connection_util import new_connection_by_tag
 
 
 @pytest.mark.asyncio
