+++ nat-lab/tests/test_teliod.py
@@ -3,7 +3,7 @@ import pytest
 from config import LIBTELIO_BINARY_PATH_DOCKER
 from contextlib import AsyncExitStack
 from helpers import setup_connections
-from utils.connection_util import ConnectionTag
+from utils.connection import ConnectionTag
 from utils.process.process import ProcessExecError
 
 TELIOD_EXEC_PATH = f"{LIBTELIO_BINARY_PATH_DOCKER}/teliod"
