+++ nat-lab/tests/test_wg_adapter.py
@@ -4,7 +4,8 @@ from datetime import datetime
 from helpers import SetupParameters, setup_environment
 from Pyro5.errors import CommunicationError  # type:ignore
 from utils.bindings import TelioAdapterType
-from utils.connection_util import ConnectionTag, new_connection_by_tag
+from utils.connection import ConnectionTag
+from utils.connection_util import new_connection_by_tag
 from utils.process import ProcessExecError
