+++ nat-lab/tests/utils/connection/__init__.py
@@ -1,3 +1,1 @@
 from .connection import *
-from .docker_connection import *
-from .ssh_connection import *
